Fix automated builds (#278)
* Switch to new bevy repo

* Update deps

* Fix clippy warnings

* Fix formatting

* Remove block macros

+++ src/inventory/slot_mapping.rs
@@ -124,7 +124,7 @@ impl SlotMapping {
                 .slots
                 .read()
                 .get_slot(x, y)
-                .map(|i| child.range[i as usize] as u16);
+                .map(|i| child.range[i as usize]);
         }
 
         None
